🚚 move etl code to subdirectory
etl is no longer at root of package
